Added ability to print arbitrary error contexts
I've been pleased with the contextual error messages printed by the
Scanner and Parser objects, so this code adds similar capabilities
within the compiler object, by extending each AST node to contain
a small amount of source positioning information (filename, line,
etc). The compiler can then reopen the source file of an offending
AST node in order to scratch-build a contextual error string.

While this feature exists in the code, it hasn't been finalized or
used anywhere. Once proper error checking is added to the compiled
code, contextual errors should be added there as well to aid in
program debugging.

+++ matte/compiler.h
@@ -72,7 +72,7 @@ struct _Compiler {
 
   /* @err: number of errors encountered by the compiler.
    */
-  int err;
+  long err;
 };
 
 /* function declarations (compiler.c): */
